vsc_parser/requirements.txt
gitpython
neo4j
networkx
python-dotenv
clang        
llvmlite      
radon         